upgrade to cass-operator 1.8.0-rc.1
Now that cas-operator is kustomized, we no longer need to store its manifests
locally. It also now has an optional dependency on cert-manager which is on
by default. This commit updates e2e tests to deploy cert-manager and adds
logic to wait for cert-manager to initialize. It is necssary to wait for the
cert-manager deployments to become ready before deploying cass-operator;
otherwise, I found that the cass-operator deployment can fail.