cass-operator v1.8.0-rc.1 (#102)
* upgrade to cass-operator 1.8.0-rc.1

Now that cas-operator is kustomized, we no longer need to store its manifests
locally. It also now has an optional dependency on cert-manager which is on
by default. This commit updates e2e tests to deploy cert-manager and adds
logic to wait for cert-manager to initialize. It is necssary to wait for the
cert-manager deployments to become ready before deploying cass-operator;
otherwise, I found that the cass-operator deployment can fail.

* run kustomize build on CRDs for integration tests

We have to pass local file system paths of CRDs to the TestEnv. We no longer
store the CassandraDatacenter CRD locally. Another problem is that
previously we were passing the CRDs without first running kustomize build on
them which means for example patches aren't applied. This commit addresses
both of these issues.

* add new crd directory which just points to the cass-operator repo

* make logging cmd output configurable, clean up command execution code a bit