Cassandra configuration [K8SSAND-822][K8SSAND-796] (#94)
* initial support for C* configs and specifying at cluster- and/or dc-level

* add integration test to check CassandraDatacenters have correct configs

This test verifies that settings specified at the cluster level are applied
correctly.

I also pass DeepCopy on the CassandraClusterTemplate that is passed into the
cassandra.Coalesce function. If we don't use a copy, we could end up using
shared references that could result in incorrect values.

* add integration test to check settings specified in dc templates

* update e2e manifests and add types with C* settings used in e2e tests

* add required field

Not sure if this should be required since it could be set in the cluster and/or
the dc templates but the focus at the moment is getting e2e tests passing.

* fix heap size property name

* fix units

Previously heap size was specified in Megabytes (MB). Now it is specified
in Mebibytes (Mi).

* add unit tests, only set thrift properties if C* version is 3.11.x

* add integration test that uses both cluster template and dc templat configs

* add AdditionalSeeds to DatacenterConfig

* clean up after rebase

* fix property names

* add file_cache_size_in_mb, row_cache_size_in_mb

I started adding auth support but decided to comment it out for now. I want to
tackle that in a follow up PR since there is some additional work involved with
the replicated secrets, and that really falls outside the scope of this PR.

* go back to running 3 C* nodes per DC and add 4th worker node for Stargate

* fix affinity for Stargate to assign it to its own worker node

* try bumping initial delay on probes to get the test passing

* bump initial delay for stargate liveness probe to see if it fixes failure

After reviewing the logs and timestamps it looks like the Stargate node in dc2
may have been restarted due to a liveness probe failure. This commit bumps
the initial delay from 90 sec to 150 sec, which hopefully is enough even in a
resource constrained environment.

* don't test deploy ingress in multi-dc test due to resource constraints

* fix validations and after test cleanup

* clean up after rebase

* create the config via json marshalling instead of trying to build it by hand

* refactor multi-cluster tests to work around resource constraints with GHA runner

* fix pod names in nodetool status checks

* fix count on nodetool status check and increase C* memory

* fix nodetool count for stargate test and increase memory

* fix rack name for new stargate test

Previously I was trying to deploy 4 worker nodes, hence specifying rack4. I
have created a new test for multi-dc with stargate. The clusters run 3 worker
nodes each.

* make Networking property configure, host networking in particular

* remove k8sContexts from manifests since it has been removed from the CRD

* update annotations

* add namespace filter to pods query

* rebase cleanup

* fix DeleteReplicatedSecrets function