Rework the reconciliation loop to deal with Stargate once all Cassandra nodes are up in all DCs.
The Stargate keyspace will be created or altered to match the desired topology only if a Stargate pod needs to be created to limit the number of alter statements being sent.
Cassandra 4.0 and Stargate are clashing for now and RF cannot be altered unless a jvm opt is added to allow rf changes during range movements (Stargate is currently not showing up in NORMAL state, which confuses Cassandra into thinking there's a topology change in progress).