implement finalizer handling for K8ssandraCluster
When deletion timestamp is set on K8ssandraCluster,
* delete all CassandraDatacenters
* delete all Stargates
* if any deletions fail, requeue
* remove the finalizer if there are no errors

Add a finalizer to the K8ssandraCluster so cleanup can be performed.

Set controller/owner reference on the ReplicatedSecret so deletes cascade to it.
We can safely set the owner reference since the ReplicatedSecret always exists
in the same namespace as the K8ssandraCluster.

Removed the managedReplicatedSecret test function from
secret_controller_test.go. The ReplicatedSecret verifcation is now done in
k8ssandracluster_controller_test.go since that is the controller where the
ReplicatedSecret is actually created.

Superuser secret verifcation is moved in
https://github.com/k8ssandra/k8ssandra-operator/pull/204 so there was no reason
to keep the managedReplicatedSecret function.