implement finalizer for K8ssandraCluster [K8SSAND-590] (#208)
* implement finalizer handling for K8ssandraCluster

When deletion timestamp is set on K8ssandraCluster,
* delete all CassandraDatacenters
* delete all Stargates
* if any deletions fail, requeue
* remove the finalizer if there are no errors

Add a finalizer to the K8ssandraCluster so cleanup can be performed.

Set controller/owner reference on the ReplicatedSecret so deletes cascade to it.
We can safely set the owner reference since the ReplicatedSecret always exists
in the same namespace as the K8ssandraCluster.

Removed the managedReplicatedSecret test function from
secret_controller_test.go. The ReplicatedSecret verifcation is now done in
k8ssandracluster_controller_test.go since that is the controller where the
ReplicatedSecret is actually created.

Superuser secret verifcation is moved in
https://github.com/k8ssandra/k8ssandra-operator/pull/204 so there was no reason
to keep the managedReplicatedSecret function.

* don't patch the status if the cluster was deleted

* rebase cleanup

* verify that the finalizer is added

* verify finalizer cleanup logic

Make sure that CassandraDatacenter and Stargate objects are deleted. There is
no need to check ReplicatedSecret. The delete will cascade for it since its
owner reference is set.

* update e2e test clean up

Remove functions that are no longer necessary. Block during cleanup until the
K8ssandraCluster is terminated.

* update changelog

* it's no longer necessary to deploy cass-operator on its own

cass-operator is deployed with k8ssandra-operator now with the recent kustomize
changes.

* Revert "update e2e test clean up"

Reverting because we have some tests that do not deploy a K8ssandraCluster so
the cleanup steps are still needed.

* don't return an error so we can continue polling

* remove commented out code

* don't check secret replication for local cluster

This commit cleans up from some merge conflict from rebase and fixes a small
issue in the HasReplicatedSecrets func. If K8sContext is empty string, that
means we are in the local cluster, so there is no need to check the replication
status.

* remove unused import