update e2e tests to support cluster-scoped deployment operator, add test
This commit includes some refactoring of the e2e setup/configuration. It
introduces the e2eTestOpts struct to configure a test for execution.

For cluster-scoped tests, you need to specify the namespace where the
K8ssandraCluster and CassandraDatacenters will be deployed. While this is a
manual step, I felt it would be less complex and easier that introduce code to
parse the manifests to determine what namespaces are needed or even some
additional templating approach.