Cluster deploy (#222)
* Add cluster wide deployment option

* use cluster-scoped cass-operator and do some clean up

* More updates to support cluster-scoped deployments

* Add Namespace field to ReplicatedTarget in support of replicating across
  namespaces
* Add kustomization patches to configure ClusterRoleBinding
* Update secret_controller.go to watch cluster-wide
* Configure caches for remote Clusters

* add kustomize dirs for cluster-scoped deployments

* update secret_controller to watch multiple namespaces, i.e., cluster scope

* Add Namespace field to ReplicationTarget
* Remove namespace predicate from queries in SecretSyncController
* Update name of ReplicatedSecret to name of K8ssandraCluster
* Update controller integration tests
* Introduce cluster-name and cluster-namespace labels
    * Both are now needed to identify the K8ssandraCluster

* rebase cleanup

* fix kustomize manifest, make sure cluster namespace label is used

Code generation added back the namespace declaration to
config/manager/manager.yaml which caused kustomize builds to fail. That has
been removed.

Add the k8ssandra.io/cluster-namespace label to all objects that have the
k8ssandra.io/cluster-name (formerly k8ssandra.io/cluster) label.

* refactor cluster scoped kustomizations and update Makefile

Updated Makefile targets to support cluster-scoped deployments.

I ran into problem we have seen before with kustomize dealing with namespaces.
I wanted to deploy both cass-operator and k8ssandra-operator in the same
namespace. Because cass-operator's cluster-scoped kustomization configures the
namespace, I am not able to change or and have problems setting it for
k8ssandra-operator. This commit introduces some changes for cass-operator
kustomizations to hopefully better deal with configuring the namespace
for a cluster-wide deployment.

* remove unused kustomization dir

* update e2e tests to support cluster-scoped deployment operator, add test

This commit includes some refactoring of the e2e setup/configuration. It
introduces the e2eTestOpts struct to configure a test for execution.

For cluster-scoped tests, you need to specify the namespace where the
K8ssandraCluster and CassandraDatacenters will be deployed. While this is a
manual step, I felt it would be less complex and easier that introduce code to
parse the manifests to determine what namespaces are needed or even some
additional templating approach.

* add temporary solution to lower mgmt api heap size for tests

* clean up after rebase from reaper controller merge

* add cluster scope e2e test to GHA workflow

* fix bad renaming

* make sure jmx changes in PodTemplateSpec are not lost

* address PR review

* Add cluster scope support to Makefile targets
* Make kustomize dir names consistent
* Add/update method for working with the cassandra container
* Add/update unit tests
* Fix code for setting labels
* Clean up e2e code

* fix e2e test cleanup logic, add tests back

* add back check that was lost during last rebase

* rebase cleanup

In addition to cleanup, modify CreateJsonConfig func to allow nil input since
it is not required to declare C* configuration in a K8ssandraCluster.

* add back tests, deploy traefik for reaper tests

* remove duplicate test

* disable reaper test that isn't working as expected

* update/fix test

* update changelog

Co-authored-by: Michael Burman <yak@iki.fi>