Check and update system replication (#244)
Store the initial replication of system keyspaces (system_auth, system_traces,
and system_distributed) in an annotation on the K8ssandraCluster. The system
replication is set via Java system properties which are passed in an env var in
the server-config init container. The values are idempontent. Changing the Java
system properties won't actually change the replication. It would however cause
an unnecessary StatefulSet update. Storing the replication in the annotation
provides an easy way to make sure that the value does not change even when
there are topology changes.

Add the updateReplicationOfSystemKeyspaces method primarily to update
replication for newly added DCs. This is needed is in support of
https://github.com/k8ssandra/k8ssandra-operator/issues/21.

Lastly, add the namespace transform back to the control-plane kustomization.
It's needed so that the k8ssandra-operator namespace is created and so that
both cass-operator and k8ssandra-operator are deployed into that namespace.