handle system keyspaces separately from reaper and stargate keyspaces
To date we have relied on setting a couple system properties to configure the
replication of system keyspaces. As I added support for managing replication
of reaper and stargate auth keyspaces, I attempted to consolidate how they are
managed. It makes sense they basically need to be managed in the same way. I
had to undo those changes though until we implement support for running repairs
after replication changes.

If we configure the replication for system_auth with client CQL calls from the
operator instead of back door in the management api, then we need to run a
repair on system_auth when a second dc is added to the cluster; otherwise,
any queries against nodes in the second dc will fail. This applies when
we are deploying a new cluster as well as when adding a dc to an existing
cluster.

This commit also updates the version of cass-operator now that the
CassandraTask API has landed in master in cass-operator.