Create first Stargate deployment with one replica, before others
This commit modifies the Stargate deployment creation logic so that
the first Stargate deployment always contains one single replica.

Moreover, this commit also modifies the Stargate controller so that
it waits until the first deployment is fully rolled out before
proceeding to creating the following ones.

This eliminates the race condition where many Stargate pods are
created at once, and all of them attempt to modify the schema
concurrently.