Fix Reaper secret reconciliation
This commit fixes Reaper secret reconciliation when
Reaper is present for some dcs, but not for all of them.

The bug hasn't been detected by e2e tests because
the multi-dc-reaper fixture uses a manually created secret.

If the Reaper secret doesn't exist when the dc is created,
we get the following error:

    Validation of user secret failed due to an error:
    Secret \"test-reaper\" not found