Update test workflows to use release image tag of operator in tests
When a PR branch's name starts with prepare-release-, e.g.
prepare-release-1.0.0-alpha.3, the release is parsed from the end. The
docker image that is built will tagged with the version string. This
assumes that the image tag in config/deployments/default/kustomization.yaml
will be updated to the same value. If not tests will fail.

I had to make some additional adjustments for kuttl tests. By default kuttl
creates the kind cluster and then runs the tests. I configured kuttl to
not create the cluster and instead create it up front so the correct image
can get loaded. Turns out there is a way to do this via kuttl, but that will
have to happen later. kuttl also starts up a separate kind cluster per test.
I have added some manual tests in the kuttl workflow to try and mimic that
behavior in an effort to get the tests passing.