store the number of nodes to rebuilt in label on the task
Changes based on PR review. There is a possibility we could end up in an
infinite requeue loop if the DC size is changed after the rebuild is started
and before it is finished. This commit avoids the problem by storing the
expected number of nodes to be rebuilt in a lable on the task.