fix logic for checking if rebuild task finished (#448)
* fix logic for checking if rebuild task finished, and use multiple nodes in e2e test

* store the number of nodes to rebuilt in label on the task

Changes based on PR review. There is a possibility we could end up in an
infinite requeue loop if the DC size is changed after the rebuild is started
and before it is finished. This commit avoids the problem by storing the
expected number of nodes to be rebuilt in a lable on the task.

* fix e2e test

* small refactoring

* remove reaper from fixture. limited resources are causing instability.