Create a standalone Medusa deployment
Allows the operator to communicate with the Medusa storage backend without requiring a Cassandra pod to be up.
Add checks in the e2e tests for the Medusa standalone pod.
The sync MedusaTask operation now uses the Medusa standalone pod instead of Cassandra pods to interact with the storage backend.
During testing, it appeared that there was no timeout on the grpc connection attempts, which could block the reconciles forever.
A new Medusa release is required for that PR to work correctly, as the Medusa standalone pod needs a config override to disable ip address resolving.