Add new SetConditionStatus to reduce duplication in the code, add restricted values for AutomatedUpdate, "always" and "once" and verify them in the webhook.
Remove tee outputting to stdout in the helm prepare script.

UpdateStatus should only delete and reset the ClusterRequiresUpdate if it was allowed to update in the first place. Also, we should listen to changes and reconcile if annotations change.

Add additional checks to the GenerationCheck test to ensure we do not touch the CassandraDatacenter unless given permission.