Remove legacy medusa purge (#1596)
* Allow running webhooks in all the dataplanes for controllers which run locally on DPs

The MedusaBackupSchedule controller for example runs on dataplanes, and has a validating webhook.
This webhook needs to be registered locally in the DP so that resources can be validated before creation.
These changes allow to properly register the webhooks across dataplanes.
This commit also introduces the ability to cleanly register controllers in the dataplanes, which should help stabilize the Medusa envtests.

* Remove the deprecated cronjob for purges in favor of MedusaBackupSchedule purges

* Fix purge schedule cleanup and add e2e tests

* Fix e2e tests

* Update changelog

* Fix e2e tests

* Fix restore check sequence in e2e tests and purge schedule name

* new fix (a good one this time?)

* Check the right thing...

* Get the backup schedules owned by the cassdc